fixed minor errors, attempting to fix numerous other erros
