--- src/main/java/edu/ucsb/cs156/organic/Interceptors/RoleUserInterceptors.java
@@ -1,19 +1,0 @@
-package edu.ucsb.cs156.organic.Interceptors;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.stereotype.Component;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Component
-public class RoleInterceptorAppConfig implements WebMvcConfigurer {
-   @Autowired
-   RoleInterceptor roleUserInterceptor;
-
-   @Override
-   public void addInterceptors(InterceptorRegistry registry) {
-      registry.addInterceptor(roleInterceptor);
-   }
-   
-}
